update5

+++ src/main/java/ComputerShop1/Controller/User/CategoryController.java
@@ -24,11 +24,11 @@ public class CategoryController extends BaseController{
 	@RequestMapping(value = "/san-pham/{id}")
 	public ModelAndView Product(@PathVariable String id) {
 		_mvShare.setViewName("user/product/category");
-		int totalData = categoryService.GetProductByIDCategory(Integer.parseInt(id)).size();
+		int totalData = categoryService.GetProductByIDCategory(Long.parseLong(id)).size();
 		PaginatesDTO paginateInfo = paginateService.GetInfoPaginates(totalData, totalProductsPage, 1);
 		_mvShare.addObject("idCategory", id);
 		_mvShare.addObject("paginateInfo", paginateInfo);
-		_mvShare.addObject("productsPaginate", categoryService.GetDataProductsPaginate(Integer.parseInt(id), paginateInfo.getStart(), totalProductsPage));
+		_mvShare.addObject("productsPaginate", categoryService.GetDataProductsPaginate(Long.parseLong(id), paginateInfo.getStart(), totalProductsPage));
 		return _mvShare;
 	}
 	
@@ -39,7 +39,7 @@ public class CategoryController extends BaseController{
 		PaginatesDTO paginateInfo = paginateService.GetInfoPaginates(totalData, totalProductsPage, Integer.parseInt(currentPage));
 		_mvShare.addObject("idCategory", id);
 		_mvShare.addObject("paginateInfo", paginateInfo);
-		_mvShare.addObject("productsPaginate", categoryService.GetDataProductsPaginate(Integer.parseInt(id) ,paginateInfo.getStart(), totalProductsPage));
+		_mvShare.addObject("productsPaginate", categoryService.GetDataProductsPaginate(Long.parseLong(id) ,paginateInfo.getStart(), totalProductsPage));
 		return _mvShare;
 	}
